% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GPG_Cycle}
\alias{GPG_Cycle}
\title{Projection onto a L1 Ball}
\usage{
GPG_Cycle(obj, lambda = 1)
}
\arguments{
\item{vec}{}

\item{radius}{}
}
\value{
projection
}
\description{
Projection onto a L1 Ball
}
\examples{
\dontrun{
   ProjectL1(1:3, 1)
}
}
