% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Projection onto a L1 Ball}
\usage{
LogLikelihood(obj)
}
\arguments{
\item{vec}{}

\item{radius}{}
}
\value{
projection
}
\description{
Projection onto a L1 Ball
}
\examples{
\dontrun{
   ProjectL1(1:3, 1)
}
}
