% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCox.R
\docType{package}
\name{MCox-package}
\alias{MCox}
\alias{MCox-package}
\title{MCox: Multi-task sparse Cox model}
\description{
Multi-task sparse Cox model.
}
\author{
\strong{Maintainer}: Simon Fontaine \email{fontaines@dms.umontreal.ca}

}
