% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCox.R
\name{MCox}
\alias{MCox}
\title{Projection onto a L1 Ball}
\usage{
MCox(data, task_index, time_index, censored_index, weight_index,
  features_indeces)
}
\arguments{
\item{vec}{}

\item{radius}{}
}
\value{
projection
}
\description{
Projection onto a L1 Ball
}
\examples{
\dontrun{
   ProjectL1(1:3, 1)
}
}
