% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ProjectL1}
\alias{ProjectL1}
\title{Projection onto a L1 Ball}
\usage{
ProjectL1(vec, radius)
}
\arguments{
\item{vec}{}

\item{radius}{}
}
\value{
projection
}
\description{
Projection onto a L1 Ball
}
\examples{
\dontrun{
   ProjectL1(1:3, 1)
}
}
